[Cherry-pick] CodeQL Fixes - Second Pass from 202202
Makes changes to comply with alerts raised by CodeQL.

A prior pass with a subset of queries used here was already done so this
is mostly addressing issues discovered by enabling a few new queries.

Most of the issues here fall into the following two categories:

1. Potential use of uninitialized pointer
2. Inconsistent integer width used in loop comparison

- [ ] Breaking change?
- Will this change break pre-existing builds or functionality without
action being taken?
  **No** - Should just be refactoring to address alerts

1. Verified CI build
2. Verified boot to EFI shell on QemuQ35Pkg

N/A

(cherry picked from commit 43e657c)

+++ UefiCpuPkg/Library/MpInitLib/PeiMpLib.c
@@ -295,7 +295,7 @@ GetWakeupBuffer (
           WakeupBufferEnd = BASE_1MB;
         }
 
-        while (WakeupBufferEnd > WakeupBufferSize) {
+        while (WakeupBufferEnd > (UINT64)WakeupBufferSize) {
           //
           // Wakeup buffer should be aligned on 4KB
           //
